The media_unittests needed to be adjusted to reflect the change made in
the dimensions used to create VideoFrames in FFmpegVideoDecodeEngine.

BUG=91506
TEST=media_unittests


Review URL: http://codereview.chromium.org/7714016

TBR=vrk@chromium.org
Review URL: http://codereview.chromium.org/7792010

git-svn-id: svn://svn.chromium.org/chrome/branches/835/src@98683 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/video/ffmpeg_video_decode_engine_unittest.cc
@@ -23,8 +23,8 @@ namespace media {
 
 static const int kWidth = 320;
 static const int kHeight = 240;
-static const int kSurfaceWidth = 522;
-static const int kSurfaceHeight = 288;
+static const size_t kSurfaceWidth = 522;
+static const size_t kSurfaceHeight = 288;
 static const AVRational kFrameRate = { 100, 1 };
 
 static void InitializeFrame(uint8_t* data, int width, AVFrame* frame) {
